Cambios sin terminar de politicas y controladores.
1. Creacion de controlador para el perfil, con su request incluido.
2. Eliminacion de vistas de search en los forms y users.
3. Mejoramiento en las busquedas realizadas para los users y forms.
4. Agregacion de politicas para usuarios.
5. Separacion de perfil de los usuarios.

Estos cambios aun no estan terminados, les faltan algunos detalles.

+++ app/Http/Controllers/UserController.php
@@ -2,18 +2,21 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\StoreUserRequest;
-use App\Http\Requests\UpdateUserRequest;
 use App\Http\Requests\UserRequest;
 use App\User;
 use App\Models\Role;
 use App\Models\Position;
 use App\Models\Career;
-use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Facades\Hash;
+use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Http\Request;
 
 class UserController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     /**
      * Muestra una lista del recurso.
      *
@@ -21,19 +24,11 @@ class UserController extends Controller
      */
     public function index()
     {
-        return view('users.index', [
-            'users' => User::paginate(8)
-        ]);
-    }
+        $this->authorize('viewAny', User::class);
 
-    /**
-     * Método que retorna el usuario que contenga un email específico. Requiere un parámetro:
-     * 
-     * @param String $email el cual es el email específico por el cual buscar al usuario.
-     */
-    public function getUser(String $email)
-    {
-        return User::where('email', $email)->get();
+        return view('users.index', [
+            'users' => User::paginate(5)
+        ]);
     }
 
     /**
@@ -43,8 +38,9 @@ class UserController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', User::class);
+
         return view('users.create', [
-            'user' => Auth::user(),
             'roles' => Role::all(),
             'positions' => Position::all(),
             'careers' => Career::all()
@@ -59,9 +55,13 @@ class UserController extends Controller
      */
     public function store(UserRequest $request)
     {
+        $this->authorize('create', User::class);
+
         User::create($request->validated());
 
-        return redirect()->route('users.index')->with('status', __('The user was created successfully.'));
+        return redirect()
+            ->route('users.index')
+            ->with('status', __('The user was created successfully.'));
     }
 
     /**
@@ -72,6 +72,8 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
+        $this->authorize('update', $user);
+
         return view('users.edit', [
             'user' => $user,
             'roles' => Role::all(),
@@ -89,50 +91,13 @@ class UserController extends Controller
      */
     public function update(UserRequest $request, User $user)
     {
+        $this->authorize('update', $user);
+
         $user->update($request->validated());
 
         return $request->routeIs('users.update') ?
             redirect()->route('users.index')->with('status', __('The user was edited successfully.')) :
             redirect()->route('users.profile')->with('status', __('Your personal information has been edited correctly.'));
-    }
-
-    /**
-     * Método que busca en la base de datos un usuario que posea un nombre o email similares a los enviados
-     * por el usuario.
-     */
-    public function search()
-    {
-        $search = request()->validate([
-            'search' => 'required'
-        ]);
-
-        $users = User::where('name', 'like', '%' . $search['search'] . '%')
-            ->orWhere('email', 'like', '%' . $search['search'] . '%')->get();
-
-        return view('users.search', [
-            'users' => $users
-        ]);
-    }
-
-    /**
-     * Método que activa o desactiva el estado de un usuario, recibe por parámetros: 
-     * 
-     * @param \App\User $user usuario al que se le va a cambiar el estado.
-     */
-    public function switchActive(User $user)
-    {
-        $user->update(['is_active' => !$user->is_active]);
-        return redirect('/users')->with('status', __('User status changed successfully'));
-    }
-
-    /**
-     * Método que redirige a la vista profile para modificación de datos del usuario en sesión
-     */
-    public function profile()
-    {
-        return view('users.profile', [
-            'user' => Auth::user(),
-        ]);
     }
 
     /**
@@ -143,12 +108,79 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        $this->authorize('delete', $user);
+
         if ($user->forms->isNotEmpty()) {
-            return redirect('/users')->with('error', __('Can not delete an user that owns a form or is register in one'));
+            return redirect()
+                ->route('users.index')
+                ->with('error', __('Can not delete an user that owns a form or is register in one'));
         }
 
         $user->delete();
 
-        return redirect()->route('users.index')->with('status', __('The user was successfully removed.'));
+        return redirect()
+            ->route('users.index')
+            ->with('status', __('The user was successfully removed.'));
+    }
+
+    /**
+     * Muestra una lista personalizada del recurso.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function search(Request $request)
+    {
+        $this->authorize('viewAny', User::class);
+
+        $request->validate([
+            'search' => 'required'
+        ]);
+
+        $search = $request->search;
+
+        $users = User::where('name', 'like', '%' . $search . '%')
+            ->orWhere('email', 'like', '%' . $search . '%')
+            ->orWhere('phone', 'like', '%' . $search . '%')
+            ->orWhereHas('career', function (Builder $query) use ($search) {
+                $query->where('name', 'like', '%' . $search . '%');
+            })
+            ->orWhereHas('position', function (Builder $query) use ($search) {
+                $query->where('name', 'like', '%' . $search . '%');
+            })
+            ->orWhereHas('role', function (Builder $query) use ($search) {
+                $query->where('label', 'like', '%' . $search . '%');
+            })
+            ->paginate(5);
+
+        return view('users.index', [
+            'users' => $users,
+            'search' => $search
+        ]);
+    }
+
+    /**
+     * Método que retorna el usuario que contenga un email específico. Requiere un parámetro:
+     * 
+     * @param String $email el cual es el email específico por el cual buscar al usuario.
+     */
+    public function getUser(String $email)
+    {
+        return User::where('email', $email)->get();
+    }
+
+    /**
+     * Método que activa o desactiva el estado de un usuario, recibe por parámetros: 
+     * 
+     * @param \App\User $user usuario al que se le va a cambiar el estado.
+     */
+    public function switchActive(User $user)
+    {
+        $this->authorize('update', $user);
+
+        $user->update(['is_active' => !$user->is_active]);
+
+        return redirect()
+            ->route('users.index')
+            ->with('status', __('User status changed successfully'));
     }
 }
